Extract shared ColumnView builder to data_grid.py
Eliminates duplication across table_panel, sql_editor, and table_view.
Any future change to column view behaviour now has a single source of truth.
