+++ src/table_panel.py
@@ -5,7 +5,9 @@ import gi
 gi.require_version('Gtk', '4.0')
 gi.require_version('Adw', '1')
 
-from gi.repository import Gtk, Adw, GLib, Pango, Gio, GObject
+from gi.repository import Gtk, Adw, GLib, Pango
+
+from data_grid import make_column_view
 
 try:
     gi.require_version('GtkSource', '5')
@@ -40,50 +42,6 @@ def _apply_scheme(buf, dark):
         buf.set_style_scheme(scheme)
 
 ROW_LIMIT = 500
-
-
-class _Row(GObject.Object):
-    __gtype_name__ = 'TuskDataRow'
-
-    def __init__(self, values):
-        super().__init__()
-        self._values = values
-
-    def get(self, i):
-        return self._values[i]
-
-
-def _make_column_view(columns, rows):
-    store = Gio.ListStore(item_type=_Row)
-    for row in rows:
-        store.append(_Row(['' if v is None else str(v) for v in row]))
-
-    col_view = Gtk.ColumnView(model=Gtk.SingleSelection(model=store))
-    col_view.set_show_row_separators(True)
-    col_view.set_show_column_separators(True)
-    col_view.set_hexpand(True)
-
-    for i, name in enumerate(columns):
-        factory = Gtk.SignalListItemFactory()
-
-        def on_setup(_factory, list_item):
-            label = Gtk.Label()
-            label.set_xalign(0)
-            label.set_ellipsize(Pango.EllipsizeMode.END)
-            label.set_max_width_chars(40)
-            list_item.set_child(label)
-
-        def on_bind(_factory, list_item, col_idx=i):
-            list_item.get_child().set_text(list_item.get_item().get(col_idx))
-
-        factory.connect('setup', on_setup)
-        factory.connect('bind', on_bind)
-
-        col = Gtk.ColumnViewColumn(title=name, factory=factory)
-        col.set_resizable(True)
-        col_view.append_column(col)
-
-    return col_view
 
 _SCHEMA_SQL = """
     SELECT column_name, data_type,
@@ -443,7 +401,7 @@ class TablePanel(Gtk.Box):
 
         # Data tab — rebuild with dynamic columns
         if data_rows:
-            self._data_scroll.set_child(_make_column_view(data_cols, data_rows))
+            self._data_scroll.set_child(make_column_view(data_cols, data_rows))
         else:
             empty = Adw.StatusPage(title='No data')
             empty.set_vexpand(True)
